small change to compile 64bit
git-svn-id: svn://svn.code.sf.net/p/flightpred/code/trunk@117 c5ce1916-e146-4858-a6ec-1f747f202df1

+++ src/common/train_svm.cpp
@@ -72,7 +72,7 @@ void train_svm::train(const string &site_name, const bgreg::date &from, const bg
         double dval;
         for(size_t j=0; j<5; ++j)
         {
-            res[i][1 + j].to(dval);
+            res[i][static_cast<unsigned int>(1 + j)].to(dval);
             if(dval < 1.0)
                 dval = 0.0;
             labelsbydate[day][j] = dval;
